requirements.txt
langchain
langchain-community
langchain_openai
openai
python-dotenv
langchain-pinecone
langgraph
grandalf
